requirements.txt
flask_cors
Flask==2.3.2
pandas==1.5.3
pandasai==0.6.7
gunicorn==20.1.0
pandas_gbq
google-auth